Update README.md
even the main developer @GreLI said it is not activly developed any more.

I know I critizied it very much, which I would not have done if you wrote it in a more neutral way.
I assume it is not purposly written that biased. If you need help or a second opinion I could assist you.